splitted logged in and out
